test: isolate timing benchmarks from default suite
- Default Jest config swept perf benchmarks into the parallel run, flaking timing thresholds
- Exclude tests/performance from the default config so benchmarks run only via test:performance
- Delete tests/integration/performance.test.js, a stale duplicate of rule-performance.test.js
- Split performance-optimization into isolated benchmarks and default-suite detection correctness
- Assert the disabled-telemetry no-op contract instead of a load-sensitive 10ms wall-clock budget
