+++ tests/performance/README.md
@@ -186,9 +186,11 @@ test('benchmark_specific_function', () => {
 
 ```text
 tests/performance/
-├── README.md                     # This file
-├── rule-performance.test.js      # Rule execution benchmarks
-└── autofix-safety.test.js        # Safety classifier benchmarks
+├── README.md                       # This file
+├── rule-performance.test.js        # Rule execution benchmarks
+├── autofix-safety.test.js          # Safety classifier benchmarks
+├── incremental-linting.test.js     # Cache warm-vs-cold benchmarks
+└── shared-utils-performance.test.js # Code-block / inline-code detection benchmarks
 ```
 
 ### rule-performance.test.js
@@ -210,6 +212,22 @@ Tests autofix safety classifier performance:
 - Individual heuristic functions (`hasCodeIndicators`, `hasCommandPattern`, etc.)
 - Batch processing performance
 - Baseline metrics for regression detection
+
+### incremental-linting.test.js
+
+Tests the linting cache's warm-vs-cold behavior:
+
+- Single-file and multi-file cache savings
+- Warm-run results match cold-run results
+
+### shared-utils-performance.test.js
+
+Tests `shared-utils` detection performance (correctness lives in
+`tests/integration/shared-utils-detection.test.js`):
+
+- Code-block detection on large documents
+- Inline-code detection and span extraction
+- Cache memory stability
 
 ## Adding new performance tests
